Use a different port and hard-code it for tests
If we reuse the existing docker container with
```
docker-compose exec embed-api yarn run mocha
```
instead of
```
docker-compose run --rm embed-api yarn run mocha
```
we need to make sure that the test server is using a different port
number than the already running server.

There are two occurences and I thought it's OK to hard-code it for now.